% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Transform_gene_ararray.R
\name{Transform_gene_array}
\alias{Transform_gene_array}
\title{Transform SPATA object into a python list (json-file)}
\usage{
Transform_gene_array(object, of_sample, tissue_pos)
}
\arguments{
\item{object}{SPATA object}

\item{of_sample}{Name of the sample of the Spata Object}

\item{tissue_pos}{File of the possition in the tissue}
}
\value{
No return but saves a json file to workdir
}
\description{
Test
}
\examples{

#Get Spata Object
library(SPATA)
input_paths = c("~/Desktop/SpatialTranscriptomics/Visium/Visium/275_T")
sample_names = c("275_T")

spata_obj <- initiateSpataObject_10X(input_paths=input_paths, 
                                    sample_names=sample_names,
                                    object_name="Obj_275_T",
                                    output_path = getwd())
                                    
tissue_pos=read.csv(paste0(input_paths,"/outs/spatial/tissue_positions_list.csv"), header=F)
names(tissue_pos)=c("barcodes", "tissue", "x_rank", "y_rank", "x", "y")
tissue_pos$barcodes=paste0(tissue_pos$barcodes, "_", sample_names) 

#Tranform data to list 
Transform_gene_array(object=spata_obj, of_sample=sample_names, tissue_pos)


}
\author{
Dieter Henrik Heiland
}
