% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Transform_gene_ararray.R
\name{Transform_gene_array}
\alias{Transform_gene_array}
\title{Transform SPATA object into a python list (json-file)}
\usage{
Transform_gene_array(object, of_sample, tissue_pos)
}
\arguments{
\item{object}{SPATA object}

\item{of_sample}{Name of the sample of the Spata Object}

\item{tissue_pos}{File of the possition in the tissue}
}
\value{
No return but saves a json file to workdir
}
\description{
Test
}
\author{
Dieter Henrik Heiland
}
